requirements.txt
psycopg2-binary==2.8.6
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
